This commit adds a safeguard to prevent a user from accidentally running the globally installed version of the script while in the development directory.
A new function, `check_for_version_mismatch`, is added and called at the start of the script. It compares the path of the executing script to the expected path of the local development version. If they are different and their content has diverged, it will print an error and exit.

Note: This check did not behave as expected in the development environment and may require further debugging.